+++ gui/src/app/RunPanel/RunPanel.tsx
@@ -1,7 +1,7 @@
 import { FunctionComponent, useCallback, useContext, useMemo } from "react";
 
 import Button from "@mui/material/Button";
-import { CompileContext } from "@SpCompilation/CompileContext";
+import { CompileContext } from "@SpSettings/Compilation/CompileContextProvider";
 import { ProjectContext } from "@SpCore/ProjectContextProvider";
 import { SamplingOpts, modelHasUnsavedChanges } from "@SpCore/ProjectDataModel";
 import StanSampler from "@SpStanSampler/StanSampler";
