+++ gui/src/app/Settings/Compilation/Constants.ts
