+++ gui/src/app/Settings/Personalization/PedanticSettingProvider.tsx
