+++ gui/src/app/Settings/SettingsWindow.tsx
@@ -1,7 +1,7 @@
-import ConfigureCompilationServerDialog from "@SpCompilation/CompilationServerDialog";
+import ConfigureCompilationServerDialog from "@SpSettings/Compilation/CompilationServerDialog";
 import CloseableDialog from "@SpComponents/CloseableDialog";
 import { FunctionComponent } from "react";
-import PersonalSettingsDialogue from "./PersonalSettingsDialogue";
+import PersonalSettingsDialogue from "./Personalization/PersonalSettingsDialogue";
 import TabWidget from "@SpComponents/TabWidget";
 
 type SettingsWindowProps = {
